Split up level reader to move the game specific data structures into separate files
